// src/pages/Products/type.ts
export type ProductDto =  {
    id: number;
    name: string;
    manufacturer: string;
    category: string;
    description: string;
    directionsForUse: string;
    warningsPrecautions: string;
    packagingSize: string;
    price: number;
    storageConditions: string;
    storageLocation: string;
    image: string;
    availability: string;  
    quantity: number;
}